Bug 1290882 - Remove 'Press back to exit' snackbar when entering fullscreen mode. r=ahunt
We decided to remove it because:
 * Modern video controls include an exit fullscreen button
 * After bug 1031519, you have to swipe down from the top of the screen to
display the soft back button so the copy does not make sense.

Note that the `MozShowFullScreenWarning` event was removed previously in the
platform in [1] so we remove the listener attachment here without replacing it.

[1]: https://hg.mozilla.org/integration/fx-team/rev/a6a5f79e630d

MozReview-Commit-ID: HwyyUkWkUUH
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: dae8eaf6a289d6f3710a3a02b9f5566ce310139e