Bug 1329109 - remove unused time and progressbar code from downloads toolbar icon. r=Paolo
This removes the arrowStyledIndicator killswitch and the associated ability for system add-ons to restore the previous design of the Downloads Indicator in the toolbar. The removal includes all the related code, styles, and strings. Based on the original patch by :rexboy.

MozReview-Commit-ID: 2OEBzPNzl0O
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: 51f1e5035fe3cefc6a52efbacd9298a34df3f100