Bug 1444327 - Bring back ability to see and copy font URLs; r=jdescottes
In the fonts panel UI prior to Firefox 60, remote font URLs used to be
displayed in full in a text input field. It made it easy to copy them.
With the redesign that happened in 60 (bug 1437548 and 1442001), getting
the URL became harder. The URL isn't visible anymore easily. There's a link
that can be clicked to load the URL in the browser, or it can also be copied
from the @font-face CSS rule code section. But that's harder.

This change adds the beginning of the URL back (with an ellipsis) and a
simple button that copies the link.

Note that the new test failed intermittently on non e10s (took too long).
This was because of a react middleware which was logging all actions, which,
in non-e10s, ended up logging StyleRuleActors, which got serialized and caused
way too much logs to be printed, slowing the test down. So the test was
disabled on non-e10s.

MozReview-Commit-ID: 2oSMoWKYhTk
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: d57cfd8c5c3928da25d7270ad33a06ccd501580e
X-Channel-Repo: releases/mozilla-beta
X-Channel-Revision: cf8705c4a1bf45b6287fbf606b4ffc205adba671
X-Channel-Repo: releases/mozilla-release
X-Channel-Revision: a46d7b315f70a022224de390ae4c1590f2d6c023