Bug 1050329 - part 2 - Add error page for h2 goaway. r=dragana,bzbarsky
This is kind of like the previous patch (where we had a not-very-friendly user experience shutting down misbehaving h2 sessions), but in this case the server has proven to us that it can speak a minimum of h2, so we don't want to just fallback. Instead, when we send a GOAWAY frame because we have detected some error on the part of the server, if it's a top-level page load, we'll show an error page explaining that the server spoke bad http/2, and the site admin(s) need to be contacted. We already did this for INADEQUATE_SECURITY (which is its own special case still), but that didn't cover all the cases.

Differential Revision: https://phabricator.services.mozilla.com/D8436

X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: 8ee3ec2214429e42c473844b86c8c1cbd0d94d15
X-Channel-Repo: releases/mozilla-beta
X-Channel-Revision: 848cf6354160ca6e2dccbeb859f64b6b9e67e4cc
X-Channel-Repo: releases/mozilla-release
X-Channel-Revision: 18d4c09e9378f1fca26c074a10195589248c88c0