Bug 1534706 - Don't focus ConfirmDialog when displaying it. r=Honza
Some people found that the ConfirmDialog was getting into their
way when typing into the console, as it was stealing the focus.
This patch fixes this by not focusing the ConfirmDialog when
we show it, so the user can still type.
This means that we now handle the dialog confirm and dismiss from
JSTerm, when the former is displayed.

Since it wasn't clear how you could close the popup, we add a close
button that makes it very obvious.

This means we can drop the key handler in the dialog as the jsterm
is always focused.
We also simply remove the feature to open the MDN link on `?` key
stroke as it's not discoverable and was the only part of the
panel where you could do such thing.

Existing tests are adapted and extended to cover the new behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D36174

X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: 607e46906b3e819c92a618904f3aa75ed844986d
X-Channel-Repo: releases/mozilla-beta
X-Channel-Revision: 9cdcc170dc1a51fc14a10d38f8aa2237d9928ba4
X-Channel-Repo: releases/mozilla-release
X-Channel-Revision: 2fb19d0466d2f61674c0af80813645cccd510593
X-Channel-Repo: releases/mozilla-esr68
X-Channel-Revision: 2fb19d0466d2f61674c0af80813645cccd510593